TowersiteThermostatControler.pde
This Arduino Sketch Monitors 2 thermostats 1 inside the building and one outside the building. If the thermostat inside the building calls for cool air the Arduino checks the second thermostat to see if the air outside the building is cool enough to open louvered vents and turn on a powerful fan to pull air through the building saving hundreds of dollars in the winter on air conditioning. Else it turns on the Air conditioning.